refactor: 소셜 로그인 요청 dto 인터페이스를 클래스로 변경 & 기존 구현체 삭제
dto의 경우, 별도의 인터페이스와 구현체가 필요하지 않다고 판단하여 기존 인터페이스를 클래스로 변경 및 파일명 변경을 진행했다. 이에 따라 기존 카카오/애플 구현체 dto를 제거하고, import 경로를 수정했다.
