--- BE/src/socialLogin/dto/kakao-login-request.dto.ts
@@ -1,7 +1,0 @@
-import { SocialLoginRequestDto } from './social-login-request.interface';
-import { IsString } from 'class-validator';
-
-export class KakaoLoginRequestDto implements SocialLoginRequestDto {
-  @IsString()
-  idToken: string;
-}
