--- BE/src/socialLogin/dto/social-login-request.interface.ts
@@ -1,3 +1,0 @@
-export interface SocialLoginRequestDto {
-  idToken: string;
-}
