+++ BE/src/socialLogin/social-login-strategy.interface.ts
@@ -1,5 +1,5 @@
-import { SocialLoginRequestDto } from './dto/social-login-request.interface';
 import { SocialWithdrawRequestDto } from './dto/social-withdraw-request.dto';
+import { SocialLoginRequestDto } from './dto/social-login-request.dto';
 
 export interface SocialLoginStrategy {
   login(
